# backend/docs/swagger.yaml
basePath: /api
definitions:
  dto.CategoryAveragePrice:
    properties:
      avg_price:
        type: number
      category:
        type: string
    type: object
  dto.CategoryResponse:
    properties:
      id:
        example: 1
        type: integer
      title:
        example: Keyboard
        type: string
    type: object
  dto.CreateProductRequest:
    properties:
      price:
        example: 150
        type: number
      title:
        example: Keyboard
        type: string
    type: object
  dto.CursorProductsResponse:
    properties:
      next_cursor:
        type: integer
      products:
        items:
          $ref: '#/definitions/dto.ProductResponse'
        type: array
    type: object
  dto.CursorUserResponse:
    properties:
      next_cursor:
        type: integer
      users:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
    type: object
  dto.ListCategoryAvaragePriceResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.CategoryAveragePrice'
        type: array
      total:
        type: integer
    type: object
  dto.ListCategoryResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.CategoryResponse'
        type: array
      total:
        type: integer
    type: object
  dto.ListOrderResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.OrderResponse'
        type: array
      total:
        type: integer
    type: object
  dto.ListProductResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.ProductResponse'
        type: array
      total:
        type: integer
    type: object
  dto.ListTotalRevenueResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.TotalRevenueResponse'
        type: array
      total:
        type: integer
    type: object
  dto.ListUserResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
      total:
        type: integer
    type: object
  dto.OrderResponse:
    properties:
      created_at:
        example: "2026-05-11 01:45:24.864701"
        type: string
      id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
    type: object
  dto.ProductResponse:
    properties:
      id:
        example: 1
        type: integer
      price:
        example: 150
        type: number
      title:
        example: Keyboard
        type: string
    type: object
  dto.TotalRevenueResponse:
    properties:
      revenue:
        type: number
      title:
        type: string
    type: object
  dto.UserResponse:
    properties:
      email:
        example: text@gmail.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: username
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Educational SQL workshop backend API
  title: Shop API
  version: "1.0"
paths:
  /categories:
    get:
      description: Returns all categories from the database
      parameters:
      - description: Number of categories to return (default 10)
        in: query
        name: limit
        type: integer
      - description: Number of categories to skip (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListCategoryResponse'
      summary: List all categories
      tags:
      - categories
  /categories/avarage-price:
    get:
      description: Returns average product price grouped by category
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListCategoryAvaragePriceResponse'
      summary: Get average price per category
      tags:
      - categories
  /orders:
    get:
      description: Returns all orders from the database
      parameters:
      - description: Number of products to return (default 10)
        in: query
        name: limit
        type: integer
      - description: Number of products to skip (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListOrderResponse'
      summary: List all orders
      tags:
      - orders
  /products:
    get:
      description: Returns all products from the database
      parameters:
      - description: Number of products to return (default 10)
        in: query
        name: limit
        type: integer
      - description: Number of products to skip (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListProductResponse'
      summary: List all products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Add a new product to the database
      parameters:
      - description: Product data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProductResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a product
      tags:
      - products
  /products/cursor:
    get:
      description: Returns products with cursor-based pagination. Pass the last product
        ID from the previous response as cursor.
      parameters:
      - description: Number of products to return (default 10)
        in: query
        name: limit
        type: integer
      - description: Last product ID from previous page
        in: query
        name: cursor
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CursorProductsResponse'
      summary: List products (cursor-based)
      tags:
      - products
  /products/revenue:
    get:
      description: Returns products sorted by total revenue (SUM of price * quantity
        per product)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListTotalRevenueResponse'
      summary: Get total revenue per product
      tags:
      - products
  /users:
    get:
      description: Returns all users from the database
      parameters:
      - description: Number of users to return (default 10)
        in: query
        name: limit
        type: integer
      - description: Number of users to skip (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListUserResponse'
      summary: List all users
      tags:
      - users
  /users/cursor:
    get:
      description: Returns users with cursor-based pagination. Pass the last user
        ID from the previous response as cursor.
      parameters:
      - description: Number of users to return (default 10)
        in: query
        name: limit
        type: integer
      - description: Last users ID from previous page
        in: query
        name: cursor
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CursorUserResponse'
      summary: List users (cursor-based)
      tags:
      - users
  /users/search:
    get:
      description: Search users by field (partial match)
      parameters:
      - description: Field to search by
        enum:
        - email
        - name
        in: query
        name: field
        required: true
        type: string
      - description: Value to search for
        in: query
        name: value
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListUserResponse'
      summary: Search user by field
      tags:
      - users
swagger: "2.0"